huggingface_space/requirements.txt
gradio>=5.0.0
numpy
Pillow
plotly
